pyproject.toml
[project]
name = "fastapi-layered-architecture"
version = "0.1.0"
description = "This is layered architecture for fastapi"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "aioboto3>=15.4.0",
    "aiohttp>=3.13.0",
    "aiomysql>=0.2.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "boto3>=1.39.12",
    "dependency-injector>=4.46.0",
    "fastapi>=0.115.12",
    "greenlet>=3.2.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "locust>=2.37.11",
    "pre-commit>=4.2.0",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.10.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "sqladmin[full]>=0.21.0",
    "sqlalchemy>=2.0.40",
    "taskiq[reload]>=0.12.1",
    "taskiq-aws>=0.4.0",
    "types-aiobotocore-s3>=2.25.0",
    "uvicorn>=0.34.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "anyio[asyncio]>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
